TEST: fix interpolate tests, leave only as import module until implemented. move TestBasicInterpolation to integration tests
